+++ packages/terra-markdown/src/terra-dev-site/test/markdown/css.test.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-import Markdown from '../../../../lib/Markdown';
-import markdown from './css.md';
-
-const CssTest = () => (
-  <Markdown src={markdown} />
-);
-
-export default CssTest;