+++ packages/terra-markdown/src/terra-dev-site/test/markdown/jsx.test.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-import Markdown from '../../../../lib/Markdown';
-import markdown from './jsx.md';
-
-const JsxTest = () => (
-  <Markdown src={markdown} />
-);
-
-export default JsxTest;