+++ packages/terra-doc-template/src/terra-dev-site/test/doc-template/common/TestComponentExample.jsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import TestComponent from './TestComponent';
-
-const TestComponentExample = () => (
-  <TestComponent text="Hey" otherText="Seeya" />
-);
-
-export default TestComponentExample;