+++ packages/terra-responsive-element/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 5.24.0 - (October 13, 2020)
+
+* Changed
+  * Replaced references to terra-doc-template with terra-docs
+
 ## 5.23.0 - (August 4, 2020)
 
 * Changed